Enable -Wdouble-promotion by default and fix related warnings (#4603)
- fix float to double implicit conversion
- add isnanf and signbitf macros, map them to existing double version when float versions are absent
- enable -Wdouble-promotion by default
- define isnan as macro for platform use core\sahred\platform\common\math.c

+++ core/iwasm/aot/aot_intrinsic.c
@@ -152,7 +152,7 @@ float64
 aot_intrinsic_fmin_f64(float64 a, float64 b)
 {
     if (isnan(a) || isnan(b))
-        return NAN;
+        return (float64)NAN;
     else if (a == 0 && a == b)
         return signbit(a) ? a : b;
     else
@@ -174,7 +174,7 @@ float64
 aot_intrinsic_fmax_f64(float64 a, float64 b)
 {
     if (isnan(a) || isnan(b))
-        return NAN;
+        return (float64)NAN;
     else if (a == 0 && a == b)
         return signbit(a) ? b : a;
     else
